+++ crates/release_plz/tests/all/helpers/mod.rs
@@ -1,2 +1,2 @@
-mod gitea;
 mod gitea_client;
+mod gitea;
